requirements.txt
Django
mysql-connector
dj-database-url
gunicorn
whitenoise
Pillow
requests
python-barcode
openpyxl
pyscard